Fix incorrect description in README.md
In the How it Works section, the SNS topic fan out description lists the incorrect regions. E.g. In normal operation, the SNS topic fans out to the active queue in the primary region, and the DR queue in the secondary region. However the documentation lists both the primary region for this, which does not match the design.